# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LexiBridge AI Local MVP API
  version: 0.8.10-pr10
  description: >
    OpenAPI contract for LexiBridge AI Local MVP, an AI-powered bilingual
    course knowledge alignment platform. This contract documents the local
    Flask + SQLite demo API and its security/error behavior.
servers:
  - url: http://127.0.0.1:5000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
    ApiSuccess:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          enum: [success]
        message:
          type: string
        request_id:
          type: string
        data:
          type: object
    ApiError:
      type: object
      required: [status, error_code, message]
      properties:
        status:
          type: string
          enum: [error]
        error_code:
          type: string
          enum:
            - AUTH_REQUIRED
            - TOKEN_EXPIRED
            - PERMISSION_DENIED
            - RESOURCE_NOT_FOUND
            - VALIDATION_ERROR
            - FILE_TOO_LARGE
            - UNSUPPORTED_FILE_TYPE
            - OCR_UNAVAILABLE
            - FORMULA_OCR_UNAVAILABLE
            - PARSING_FAILED
            - QUOTA_EXCEEDED
            - AI_PROVIDER_FAILED
            - AI_PROVIDER_NOT_CONFIGURED
            - AI_INVALID_RESPONSE
            - PDF_FONT_UNAVAILABLE
            - TOO_MANY_REQUESTS
            - INTERNAL_ERROR
        message:
          type: string
        request_id:
          type: string
        details:
          type: object
          description: Error details may include audit_error_code for review policy, permission, and validation failures.
...
  responses:
    Error400:
      description: Validation error
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
    Error401:
      description: Authentication required or token expired
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
    Error403:
      description: Permission denied
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
    Error404:
      description: Resource not found
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
...
paths:
...
  /api/auth/login:
    post:
      security: []
      summary: Login and receive bearer token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: {type: string}
                password: {type: string}
      responses:
        "200": {description: Token and user, content: {application/json: {schema: {$ref: "#/components/schemas/ApiSuccess"}}}}
        "401": {$ref: "#/components/responses/Error401"}
...
  /api/alignment/verify:
    post:
      summary: Run alignment verification through governed providers
      description: >
        Runs mock, fake, replay, or disabled external alignment providers
        through the provider contract, governance gate, parser, and audit
        chain. Provider output never auto-approves ConceptAlignmentCards and
        never writes ConceptAlignmentCard.confidence_score.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                card_uid: {type: string}
                provider: {type: string}
                attach_to_card: {type: boolean}
                english_term: {type: string}
                chinese_term: {type: string}
                course: {type: string}
                chapter: {type: string}
                english_evidence: {type: array, items: {type: object}}
                chinese_evidence: {type: array, items: {type: object}}
                fake_response_type: {type: string}
                replay_response_type: {type: string}
      responses:
        "200": {description: Verification run result, content: {application/json: {schema: {$ref: "#/components/schemas/ApiSuccess"}}}}
        "400": {$ref: "#/components/responses/Error400"}
        "403": {$ref: "#/components/responses/Error403"}
        "404": {$ref: "#/components/responses/Error404"}
...